API/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
graphene-django>=2.1,<3
python-decouple==3.5
djangorestframework==3.11.2
Pillow==2.6.1